refactor(): moved out feeds model to seperate module
